--- a/postprocess/functions.go
+++ b/postprocess/functions.go
@@ -2,6 +2,8 @@ package postprocess
 
 import (
 	"bytes"
+	"math"
+	"reflect"
 	"regexp"
 	"strings"
 
@@ -21,39 +23,49 @@ import (
 var functions = map[string]func(*CompileContext, *Config) (Function, error){
 	// functions defined in tilezen/vector-datasource.
 	// nil values have not been implemented.
-	"numeric_min_filter":              compileNumericMinFilter,
-	"road_networks":                   compileRoadNetworks,
-	"build_fence":                     nil,
-	"drop_properties":                 nil,
-	"csv_match_properties":            compileCSVMatchProperties,
-	"exterior_boundaries":             nil,
-	"drop_features_mz_min_pixels":     nil,
-	"overlap":                         nil,
-	"admin_boundaries":                nil,
-	"handle_label_placement":          compileHandleLabelPlacement,
-	"remove_duplicate_features":       compileRemoveDuplicateFeatures,
-	"drop_features_where":             compileDropFeaturesWhere,
-	"merge_line_features":             nil,
-	"merge_building_features":         nil,
-	"merge_polygon_features":          nil,
-	"generate_address_points":         nil,
-	"merge_duplicate_stations":        nil,
-	"normalize_station_properties":    nil,
-	"rank_features":                   nil,
-	"update_parenthetical_properties": nil,
-	"keep_n_features":                 nil,
-	"drop_properties_with_prefix":     nil,
-	"drop_small_inners":               nil,
-	"simplify_and_clip":               nil,
-	"intercut":                        nil,
-	"simplify_layer":                  nil,
-	"backfill_from_other_layer":       compileBackfillFromOtherLayers,
-	"buildings_unify":                 nil,
-	"palettize_colours":               nil,
-	"point_in_country_logic":          nil,
-	"tags_set_ne_min_max_zoom":        nil,
-	"drop_layer":                      nil,
-	"max_zoom_filter":                 nil,
+	"numeric_min_filter":                 compileNumericMinFilter,
+	"road_networks":                      compileRoadNetworks,
+	"build_fence":                        nil,
+	"drop_properties":                    nil, // TODO
+	"csv_match_properties":               compileCSVMatchProperties,
+	"exterior_boundaries":                nil,
+	"drop_features_mz_min_pixels":        nil,
+	"overlap":                            nil, // look into
+	"admin_boundaries":                   nil,
+	"apply_disputed_boundary_viewpoints": nil,
+	"drop_names_on_short_boundaries":     nil,
+	"handle_label_placement":             compileHandleLabelPlacement,
+	"remove_duplicate_features":          compileRemoveDuplicateFeatures,
+	"drop_features_where":                compileDropFeaturesWhere,
+	"merge_line_features":                nil,
+	"merge_building_features":            nil,
+	"merge_polygon_features":             nil,
+	"generate_address_points":            nil,
+	"merge_duplicate_stations":           nil,
+	"normalize_station_properties":       nil,
+	"rank_features":                      nil,
+	"update_parenthetical_properties":    nil,
+	"keep_n_features":                    nil,
+	"drop_properties_with_prefix":        nil,
+	"drop_small_inners":                  nil,
+	"simplify_and_clip":                  nil,
+	"intercut":                           nil,
+	"simplify_layer":                     nil,
+	"backfill_from_other_layer":          compileBackfillFromOtherLayers,
+	"buildings_unify":                    nil,
+	"palettize_colours":                  nil,
+	"point_in_country_logic":             nil,
+	"tags_set_ne_min_max_zoom":           nil,
+	"drop_layer":                         nil, // drops admin_areas layer, which we completely ignore
+	"max_zoom_filter":                    nil, // Natural Earth layers, drop if out of range
+	"min_zoom_filter":                    nil,
+	"update_min_zoom":                    nil,              // TODO
+	"remap":                              nil,              // TODO
+	"drop_names":                         compileDropNames, // only hits zoom 13 on landuse
+	"whitelist":                          nil,              // remove road properites at lower zooms
+	"quantize_height":                    compileQuantizeHeight,
+	"clamp_min_zoom":                     nil, // TODO
+	"add_collision_rank":                 nil, // TODO
 }
 
 var (
@@ -510,12 +522,12 @@ func (f *dropFeaturesWhere) Eval(ctx *Context, layers map[string]*geojson.Featur
 		return
 	}
 
-	at := 0
 	layer := layers[f.Layer]
 	if layer == nil {
 		return
 	}
 
+	at := 0
 	for _, feature := range layer.Features {
 		ctx.fctx = filter.NewContextFromProperties(ctx.fctx, feature.Properties)
 		if f.Condition.Eval(ctx.fctx) {
@@ -533,7 +545,15 @@ func compileDropFeaturesWhere(ctx *CompileContext, c *Config) (Function, error) 
 	f := &dropFeaturesWhere{}
 
 	f.Layer = c.Params["source_layer"].(string)
-	f.StartZoom = float64(c.Params["start_zoom"].(int))
+	zs, ok := c.Params["start_zoom"]
+	if ok {
+		z, ok := zs.(int)
+		if !ok {
+			return nil, errors.New("drop_features_where: start_zoom must be an integer")
+		}
+
+		f.StartZoom = float64(z)
+	}
 
 	cond, err := filter.CompileCondition(c.Params["where"])
 	if err != nil {
@@ -541,6 +561,190 @@ func compileDropFeaturesWhere(ctx *CompileContext, c *Config) (Function, error) 
 	}
 
 	f.Condition = cond
+	return f, nil
+}
+
+// keyIsName returns true if this key looks like a name.
+// This isn't as simple as testing if key == 'name', as there are alternative
+// name-like tags such as 'official_name', translated names such as 'name:en',
+// and left/right names for boundaries. This function aims to match all of
+// those variants.
+func keyIsName(key string) bool {
+	// simplest and most common case first
+	if key == "name" {
+		return true
+	}
+
+	// translations next
+	if strings.HasPrefix(key, "name:") {
+		return true
+	}
+
+	// then any of the alternative forms of name
+	tagTameAlternates := []string{
+		"int_name",
+		"loc_name",
+		"nat_name",
+		"official_name",
+		"old_name",
+		"reg_name",
+		"short_name",
+		"name_left",
+		"name_right",
+		"name:short",
+	}
+
+	for _, alt := range tagTameAlternates {
+		if strings.HasPrefix(key, alt) {
+			return true
+		}
+	}
+
+	return false
+}
+
+type dropNames struct {
+	Layer     string
+	StartZoom float64
+	EndZoom   float64
+}
+
+func (f *dropNames) Eval(ctx *Context, layers map[string]*geojson.FeatureCollection) {
+	if ctx.Zoom < f.StartZoom || f.EndZoom < ctx.Zoom {
+		return
+	}
+
+	layer := layers[f.Layer]
+	if layer == nil {
+		return
+	}
+
+	for _, feature := range layer.Features {
+		for k := range feature.Properties {
+			if keyIsName(k) {
+				delete(feature.Properties, k)
+			}
+		}
+	}
+}
+
+func compileDropNames(ctx *CompileContext, c *Config) (Function, error) {
+	f := &dropNames{EndZoom: 50}
+
+	f.Layer = c.Params["source_layer"].(string)
+	zs, ok := c.Params["start_zoom"]
+	if ok {
+		z, ok := zs.(int)
+		if !ok {
+			return nil, errors.New("drop_nams: start_zoom must be an integer")
+		}
+
+		f.StartZoom = float64(z)
+	}
+
+	ze, ok := c.Params["end_zoom"]
+	if ok {
+		z, ok := ze.(int)
+		if !ok {
+			return nil, errors.New("drop_names: end_zoom must be an integer")
+		}
+
+		f.EndZoom = float64(z)
+	}
+
+	return f, nil
+}
+
+type quantizeHeight struct {
+	Layer     string
+	StartZoom float64
+	EndZoom   float64
+}
+
+func quantize(val, step float64) float64 {
+	// special case: if val is very small, we don't want it rounding to zero, so
+	// round the smallest values up to the first step.
+	if val < step {
+		return math.Floor(step)
+	}
+
+	return math.Floor(step * math.Round(val/step))
+}
+
+func (f *quantizeHeight) Eval(ctx *Context, layers map[string]*geojson.FeatureCollection) {
+	if ctx.Zoom < f.StartZoom || f.EndZoom < ctx.Zoom {
+		return
+	}
+
+	layer := layers[f.Layer]
+	if layer == nil {
+		return
+	}
+
+	if ctx.Zoom == 13 {
+		for _, feature := range layer.Features {
+			if height, ok := feature.Properties["height"].(float64); ok {
+				feature.Properties["height"] = quantize(height, 20)
+			}
+		}
+	}
+
+	if ctx.Zoom == 14 {
+		for _, feature := range layer.Features {
+			if height, ok := feature.Properties["height"].(float64); ok {
+				feature.Properties["height"] = quantize(height, 10)
+			}
+		}
+	}
+
+	if ctx.Zoom == 15 {
+		for _, feature := range layer.Features {
+			if height, ok := feature.Properties["height"].(float64); ok {
+				feature.Properties["height"] = quantize(height, 10)
+			}
+		}
+	}
+}
+
+func compileQuantizeHeight(ctx *CompileContext, c *Config) (Function, error) {
+	f := &quantizeHeight{EndZoom: 50}
+
+	f.Layer = c.Params["source_layer"].(string)
+	zs, ok := c.Params["start_zoom"]
+	if ok {
+		z, ok := zs.(int)
+		if !ok {
+			return nil, errors.New("quantize_height: start_zoom must be an integer")
+		}
+
+		f.StartZoom = float64(z)
+	}
+
+	ze, ok := c.Params["end_zoom"]
+	if ok {
+		z, ok := ze.(int)
+		if !ok {
+			return nil, errors.New("quantize_height: end_zoom must be an integer")
+		}
+
+		f.EndZoom = float64(z)
+	}
+
+	quantize, ok := c.Params["quantize"].(map[interface{}]interface{})
+	if !ok {
+		return nil, errors.New("quantize_height: is required")
+	}
+
+	// behavior is hard coded so will need to do something different if things change.
+	expected := map[interface{}]interface{}{
+		13: "vectordatasource.transform.quantize_height_round_nearest_20_meters",
+		14: "vectordatasource.transform.quantize_height_round_nearest_10_meters",
+		15: "vectordatasource.transform.quantize_height_round_nearest_10_meters",
+	}
+	if !reflect.DeepEqual(quantize, expected) {
+		return nil, errors.New("quantize_height: quanize has changed")
+	}
+
 	return f, nil
 }