The configs and supporting code are from the v1.5.0ish days. This PR is a work in progress to bring it up to [v1.8.0](https://github.com/tilezen/vector-datasource/releases/tag/v1.8.0)

These filter functions were added for some POI filters: 
* `tz_estimate_parking_capacity`
* `tz_looks_like_service_area`
* `tz_looks_like_rest_area`

Added these transforms:
* `population_rank`
* `major_airport_detector`
* `add_vehicle_restrictions`
* `pois_direction_int`

Added a lot of postprocessors that might be useful:
* `remap` - clean up some field values at zoom 13
* `drop_names` - used by landuse at zoom 13
* `whitelist` - approved values for certain fields, like road types.
* `quantize_height` - building heights
* `clamp_min_zoom` - keep features in higher zoom if they have a low scale_rank
* `add_collision_rank` - big one to sort label placement
* `update_parenthetical_properties` - removes feature if `(closed)` or similar at the end of the name

**Benchmarks**
```
benchmark                           old ns/op     new ns/op     delta
BenchmarkFullTile-12                6680182       7403002       +10.82%
BenchmarkProcessGeoJSON-12          4874742       5585685       +14.58%

benchmark                           old allocs     new allocs   delta
BenchmarkFullTile-12                21920          21983        +0.29%
BenchmarkProcessGeoJSON-12          11536          11599        +0.55%

benchmark                           old bytes     new bytes     delta
BenchmarkFullTile-12                2625968       2830462       +7.79%
BenchmarkProcessGeoJSON-12          1178621       1381118       +17.18%
```

Not sure why the runtime and bytes have increased. A lot is different.